feat(frontend/copilot): restyle suggestion pills and extract file drop hook
Replace bordered pill buttons with a vertical list layout using Sparkle
icons for better readability. Extract drag-and-drop file handling from
CopilotPage into a dedicated useCopilotFileDrop hook.
